+++ src/libecalc/presentation/yaml/mappers/process_simulation_mapper.py
@@ -41,10 +41,10 @@ from libecalc.presentation.yaml.yaml_types.components.yaml_process_system import
     YamlProcessPipeline,
     YamlProcessSimulation,
 )
-from libecalc.presentation.yaml.yaml_types.components.yaml_process_units import YamlCompressor, YamlCompressorModelChart
 from libecalc.presentation.yaml.yaml_types.models import YamlFluidModel
 from libecalc.presentation.yaml.yaml_types.models.yaml_compressor_stages import YamlControlMarginUnits
 from libecalc.presentation.yaml.yaml_types.models.yaml_fluid import YamlCompositionFluidModel, YamlPredefinedFluidModel
+from libecalc.presentation.yaml.yaml_types.process.yaml_process_units import YamlCompressor, YamlCompressorModelChart
 from libecalc.presentation.yaml.yaml_types.streams.yaml_inlet_stream import YamlInletStream, YamlInletStreamRate
 from libecalc.presentation.yaml.yaml_types.yaml_data_or_file import YamlFile
 from libecalc.process.fluid_stream.fluid_model import FluidModel
