+++ src/libecalc/presentation/yaml/yaml_models/pyyaml_yaml_model.py
@@ -32,10 +32,10 @@ from libecalc.presentation.yaml.yaml_types.components.yaml_process_system import
     YamlProcessSimulation,
     YamlProcessSystem,
 )
-from libecalc.presentation.yaml.yaml_types.components.yaml_process_units import YamlProcessUnit
 from libecalc.presentation.yaml.yaml_types.facility_model.yaml_facility_model import YamlFacilityModel
 from libecalc.presentation.yaml.yaml_types.fuel_type.yaml_fuel_type import YamlFuelType
 from libecalc.presentation.yaml.yaml_types.models import YamlConsumerModel, YamlFluidModel
+from libecalc.presentation.yaml.yaml_types.process.yaml_process_units import YamlProcessUnit
 from libecalc.presentation.yaml.yaml_types.streams.yaml_inlet_stream import YamlInletStream
 from libecalc.presentation.yaml.yaml_types.time_series.yaml_time_series import YamlTimeSeriesCollection
 from libecalc.presentation.yaml.yaml_types.yaml_default_datetime import YamlDefaultDatetime
